🔻 Use float precision for untyped damage scalars

+++ src/Bearded.TD/Game/Simulation/Damage/UntypedDamage.cs
@@ -24,18 +24,18 @@ readonly struct UntypedDamage
     public static UntypedDamage operator *(int scalar, UntypedDamage amount) =>
         new(scalar * amount.Amount);
 
-    public static UntypedDamage operator *(double scalar, UntypedDamage amount) =>
+    public static UntypedDamage operator *(float scalar, UntypedDamage amount) =>
         new(((int) (scalar * amount.Amount.NumericValue)).HitPoints());
 
     public static UntypedDamage operator *(UntypedDamage amount, int scalar) =>
         new(scalar * amount.Amount);
 
-    public static UntypedDamage operator *(UntypedDamage amount, double scalar) =>
+    public static UntypedDamage operator *(UntypedDamage amount, float scalar) =>
         new(((int) (scalar * amount.Amount.NumericValue)).HitPoints());
 
     public static UntypedDamage operator /(UntypedDamage amount, int scalar) =>
         new((amount.Amount.NumericValue / scalar).HitPoints());
 
-    public static UntypedDamage operator /(UntypedDamage amount, double scalar) =>
+    public static UntypedDamage operator /(UntypedDamage amount, float scalar) =>
         new(((int) (amount.Amount.NumericValue / scalar)).HitPoints());
 }
